api get base

+++ src/Pizza4Ps.PizzaService.Application/Models/PaginatedResult.cs
@@ -3,9 +3,9 @@
     public class PaginatedResult<T>
     {
         public List<T> Items { get; }
-        public int TotalCount { get; }
+        public long TotalCount { get; }
 
-        public PaginatedResult(List<T> items, int totalCount)
+        public PaginatedResult(List<T> items, long totalCount)
         {
             Items = items;
             TotalCount = totalCount;
